Merge 'feat-crossplatform' into 'main'

+++ shared/src/network/Client.cpp
@@ -113,7 +113,7 @@ bool network::NetworkClient::sendInputPacket(const struct InputPacket &packet)
     newPacket.header.packetId = lastPacketId;
 
     auto now = std::chrono::steady_clock::now();
-    newPacket.header.timestamp = std::chrono::duration_cast<std::chrono::milliseconds>(now.time_since_epoch()).count();
+    newPacket.header.timestamp = static_cast<uint32_t>(std::chrono::duration_cast<std::chrono::milliseconds>(now.time_since_epoch()).count());
 
     newPacket.header.type = PacketType::InputPacket;
 
@@ -138,7 +138,7 @@ bool network::NetworkClient::sendLobbyPacket(const struct LobbyActionPacket &pac
     newPacket.header.packetId = lastPacketId;
 
     auto now = std::chrono::steady_clock::now();
-    newPacket.header.timestamp = std::chrono::duration_cast<std::chrono::milliseconds>(now.time_since_epoch()).count();
+    newPacket.header.timestamp = static_cast<uint32_t>(std::chrono::duration_cast<std::chrono::milliseconds>(now.time_since_epoch()).count());
 
     newPacket.header.type = PacketType::LobbyActionPacket;
